+++ core/src/main/java/com/atomguard/manager/ConfigManager.java
@@ -62,7 +62,7 @@ public class ConfigManager {
     }
 
     private void checkConfigVersion() {
-        String currentVersion = "2.0.2";
+        String currentVersion = "2.0.3";
         String configVersion = config.getString("config-version", "1.0.0");
 
         if (!configVersion.equals(currentVersion)) {